requirements.txt
# Sort by alphabet
bs4
cssselect
ffmpeg-python
google-api-python-client
google-auth
google-auth-httplib2
lxml
matplotlib
numpy
pafy
pydub
pylint
pytest
pytube
requests
scipy
youtube-dl
graphviz